Add files via upload
The following script was written by request of an acquaintance of mine who is into trading and intra-day trading in particular.
The reason for it being written is that this day-trader had observed several "gaps" in leading stock indicies, and wanted to get a more quantitative and analytical understanding of these types of price action. The script is written purely by myself, and might contain inconvenient lines of code, that can be written in a more clean and proper manner, however it gets the job done in the end.